refactoring add folder ssr
