+++ index.html
@@ -8,6 +8,6 @@
 
   <body>
     <div id="root" class="App container">ssr-outlet</div>
-    <script type="module" src="./src/entry-client.tsx"></script>
+    <script type="module" src="/src/ssr/entry-client.tsx"></script>
   </body>
 </html>
